+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Infrastructure maintenance:
   * Included "Current status" section in README.md. [requested by CaverBruce]
   * Included "Choose THConfig and run Therion" and "Run Therion" buttons in both MapiaHome and TH2FileEdit pages.
+  * Moving business logic of running therion from MPRunTherionDialogWidget to new MPTherionRunner.
 
 ## 0.2.36 - 2026-02-24 - The [Danaids](https://en.wikipedia.org/wiki/Danaids) release
 * New features: